[BUGFIX] Statistics module uses different docheader menu code
This was the cause for non-translated menu headlines.
With this fix all controllers use the same docheader menu code.

(Moved the code to AbstractBackendController, controllers without header menus skip the function completely)
